+++ one/remote/globus.py
@@ -3,21 +3,25 @@ import re
 import sys
 import logging
 from pathlib import Path
-
-from iblutil.io import params as iopar
-from one.alf.spec import is_uuid, is_uuid_string
-from one.api import ONE
+import warnings
 
 import globus_sdk
 from globus_sdk import TransferAPIError, GlobusAPIError, NetworkError, GlobusTimeoutError, \
     GlobusConnectionError, GlobusConnectionTimeoutError
+from iblutil.io import params as iopar
+
+from one.alf.spec import is_uuid, is_uuid_string
+from one.alf.io import remove_uuid_file
+import one.params
+from one.webclient import AlyxClient
+from .base import DownloadClient, load_client_params, save_client_params
 
 _logger = logging.getLogger(__name__)
-_PAR_ID_STR = 'globus'
+CLIENT_KEY = 'globus'
 DEFAULT_PAR = {'GLOBUS_CLIENT_ID': None, 'local_endpoint': None, 'local_path': None}
 
 
-def setup():
+def setup(par_id=None):
     """
     Sets up Globus as a backend for ONE functions.
 
@@ -26,88 +30,97 @@ def setup():
        (https://globus-sdk-python.readthedocs.io/en/stable/tutorial.html).
     2. Set up Global Connect on your local device (https://www.globus.org/globus-connect-personal).
     3. Register your local device as an Endpoint in your Globus Client (https://app.globus.org/).
+
+    Parameters
+    ----------
+    par_id : str
+        Parameter profile name to set up e.g. 'default', 'admin'.
+
     """
 
-    print("Setting up Globus parameter file. See docstring for help.")
-    par_id = input("Enter name for this client or press Enter to keep value 'default': ")
-    par_id = _PAR_ID_STR + '/default' if par_id == '' else _PAR_ID_STR + f'/{par_id.strip(".")}'
+    print('Setting up Globus parameter file. See docstring for help.')
+    if not par_id:
+        default_par_id = 'default'
+        par_id = input(
+            f'Enter name for this client or press Enter to keep value "{default_par_id}": '
+        )
+        par_id = par_id.strip() or default_par_id
 
     # Read existing globus params if present
-    pars = iopar.read(par_id, DEFAULT_PAR)
+    globus_pars = iopar.as_dict(load_client_params(CLIENT_KEY, assert_present=False) or {})
+    pars = {**DEFAULT_PAR, **globus_pars.get(par_id, {})}
 
     # Set GLOBUS_CLIENT_ID
-    current_id = getattr(pars, 'GLOBUS_CLIENT_ID')
+    current_id = pars['GLOBUS_CLIENT_ID']
     if current_id:
-        new_id = input(f'Found Globus client ID in parameter file ({current_id}). Press Enter to '
-                       f'keep it, or enter a new ID here: ').strip()
-        new_id = current_id if not new_id else new_id
+        prompt = (f'Found Globus client ID in parameter file ({current_id}). '
+                  'Press Enter to keep it, or enter a new ID here: ')
+        pars['GLOBUS_CLIENT_ID'] = input(prompt).strip() or current_id
     else:
         new_id = input(f'Please enter the Globus client ID: ').strip()
-    pars = pars.set('GLOBUS_CLIENT_ID', new_id)
+        if not new_id:
+            raise ValueError('Globus client ID is a required field')
+        pars['GLOBUS_CLIENT_ID'] = new_id
 
     # Find and set local ID
+    message = 'Please enter the local endpoint ID'
     try:
-        local_endpoint = get_local_endpoint_id()
+        default_endpoint = pars['local_endpoint'] or get_local_endpoint_id()
+        message += f' (default: {default_endpoint})'
     except AssertionError:
-        try:
-            local_endpoint = getattr(pars, 'local_endpoint')
-            print(f"Found local endpoint ID in parameter file ({local_endpoint}).")
-        except AttributeError:
-            local_endpoint = input(
-                "Cannot find local endpoint ID. Beware that this might mean that Globus Connect "
-                "is not set up properly. You can enter the local endpoint ID manually here: ")
-    pars = pars.set('local_endpoint', local_endpoint)
+        default_endpoint = ''
+        warnings.warn(
+            'Cannot find local endpoint ID. Beware that this might mean that Globus Connect '
+            'is not set up properly.')
+    pars['local_endpoint'] = input(message + ':').strip() or default_endpoint
+    if not pars['local_endpoint']:
+        raise ValueError('Globus local endpoint ID is a required field')
 
     # Check for local path
-    try:
-        local_path = get_local_endpoint_path()
-    except AssertionError:
-        try:
-            local_path = getattr(pars, 'local_path')
-            print(f"Found local endpoint path in parameter file ({local_path}).")
-        except AttributeError:
-            local_path = input(
-                "Cannot find local endpoint path accessible by Globys. Beware that this might mean"
-                "that Globus Connect is not set up properly. Press Enter to leave this field "
-                "empty, or enter path here: ")
-
-    pars = pars.set('local_path', local_path)
+    message = 'Please enter the local endpoint path'
+    local_path = pars['local_path'] or one.params.get(silent=True).CACHE_DIR
+    message += f' (default: {local_path})'
+    pars['local_path'] = input(message + ':').strip() or local_path
 
     # Log in manually and get refresh token to avoid having to login repeatedly
-    client = globus_sdk.NativeAppAuthClient(pars.GLOBUS_CLIENT_ID)
+    client = globus_sdk.NativeAppAuthClient(pars['GLOBUS_CLIENT_ID'])
     client.oauth2_start_flow(refresh_tokens=True)
     authorize_url = client.oauth2_get_authorize_url()
     print('To get a new token, go to this URL and login: {0}'.format(authorize_url))
     auth_code = input('Enter the code you get after login here (press "c" to cancel): ').strip()
-    if auth_code in ['c', 'C', '']:
-        return
+    if auth_code and auth_code.lower() != 'c':
+        token_response = client.oauth2_exchange_code_for_tokens(auth_code)
+        globus_transfer_data = token_response.by_resource_server['transfer.api.globus.org']
+        for par in ['refresh_token', 'access_token', 'expires_at_seconds']:
+            pars[par] = globus_transfer_data[par]
 
-    token_response = client.oauth2_exchange_code_for_tokens(auth_code)
-    globus_transfer_data = token_response.by_resource_server['transfer.api.globus.org']
-    for par in ['refresh_token', 'access_token', 'expires_at_seconds']:
-        pars = pars.set(par, globus_transfer_data[par])
-
-    iopar.write(par_id, pars)
-    print("Finished setup.")
+    globus_pars[par_id] = pars
+    save_client_params(globus_pars, client_key=CLIENT_KEY)
+    print('Finished setup.')
 
 
 def create_globus_client(client_name='default'):
     """
     Creates a Globus transfer client based on existing parameter file.
 
-    :param client_name: str, defines the parameter file to use (~/.globus/.client_name)
-    :return: Globus transfer client
+    Parameters
+    ----------
+    client_name : str
+        Defines the parameter name to use (globus.client_name), e.g. 'default', 'admin'.
+
+    Returns
+    -------
+    globus_sdk.TransferClient
+        Globus transfer client instance
     """
-    par_id = _PAR_ID_STR + f'/{client_name}'
     try:
-        globus_pars = iopar.read(par_id)
-    except FileNotFoundError as err:
-        _logger.error(f"{str(err)}. Choose a different client name or run one.globus.setup() "
-                      f"first\n")
-        return
+        globus_pars = load_client_params(f'{CLIENT_KEY}.{client_name}')
+    except (AttributeError, FileNotFoundError):
+        setup(client_name)
+        globus_pars = load_client_params(f'{CLIENT_KEY}.{client_name}', assert_present=False) or {}
     required_fields = {'refresh_token', 'GLOBUS_CLIENT_ID'}
-    if not (globus_pars and required_fields.issubset(globus_pars.as_dict())):
-        raise ValueError("No token in client parameter file. Run one.globus.setup first")
+    if not (globus_pars and required_fields.issubset(iopar.as_dict(globus_pars))):
+        raise ValueError('No token in client parameter file. Run one.globus.setup first')
     client = globus_sdk.NativeAppAuthClient(globus_pars.GLOBUS_CLIENT_ID)
     client.oauth2_start_flow(refresh_tokens=True)
     authorizer = globus_sdk.RefreshTokenAuthorizer(globus_pars.refresh_token, client)
@@ -117,104 +130,98 @@ def create_globus_client(client_name='default'):
 def get_local_endpoint_id():
     """
     Extracts the ID of the local Globus Connect endpoint.
-    :return: str, local endpoint ID
-    """
-    msg = ("Cannot find local endpoint ID, check if Globus Connect is set up correctly, "
-           "{} exists and contains a UUID.")
-    if sys.platform == 'win32' or sys.platform == 'cygwin':
-        id_path = Path(os.environ['LOCALAPPDATA']).joinpath("Globus Connect")
-    else:
-        id_path = Path.home().joinpath(".globusonline", "lta")
 
-    id_file = id_path.joinpath("client-id.txt")
-    assert (id_file.exists()), msg.format(id_file)
+    Returns
+    -------
+    str
+        The local Globus endpoint ID
+    """
+    msg = ('Cannot find local endpoint ID, check if Globus Connect is set up correctly, '
+           '{} exists and contains a UUID.')
+    if sys.platform == 'win32' or sys.platform == 'cygwin':
+        id_path = Path(os.environ['LOCALAPPDATA']).joinpath('Globus Connect')
+    else:
+        id_path = Path.home().joinpath('.globusonline', 'lta')
+
+    id_file = id_path.joinpath('client-id.txt')
+    assert id_file.exists(), msg.format(id_file)
     local_id = id_file.read_text().strip()
-    assert (isinstance(local_id, str)), msg.format(id_file)
-    print(f"Found local endpoint ID in Globus Connect settings {local_id}")
+    assert isinstance(local_id, str), msg.format(id_file)
+    print(f'Found local endpoint ID in Globus Connect settings {local_id}')
     return local_id
 
 
-def get_local_endpoint_path():
+def get_local_endpoint_paths():
     """
-    Extracts the local endpoint path accessible by Globus Connect.
-    :return: str, local_path
+    Extracts the local endpoint paths accessible by Globus Connect.  NB: This is only supported
+    on Linux.
+
+    Returns
+    -------
+    list of pathlib.Path
+        Local endpoint paths set in Globus Connect.
     """
-    msg = ("Cannot find local endpoint path, check if Globus Connect is set up correctly, "
-           "{} exists and contains a valid path.")
     if sys.platform == 'win32':
-        local_path_input = input(f'On windows the local globus path needs to be entered manually. Press Enter to use the same'
-                                 f' path as the ONE cache dir or enter a path here: ').strip()
-        local_path = ONE().cache_dir if not local_path_input else Path(local_path_input)
+        print('On windows the local Globus path needs to be entered manually')
+        return []
     else:
-        path_file = Path.home().joinpath(".globusonline", "lta", "config-paths")
+        path_file = Path.home().joinpath('.globusonline', 'lta', 'config-paths')
         if path_file.exists():
-            local_path = Path(path_file.read_text().strip().split(',')[0])
-
-    if local_path.exists():
-        print(f"Found local endpoint path in Globus Connect settings {local_path}")
-        return str(local_path)
-    _logger.warning(msg.format(path_file))
-    return None
-
-
-def get_endpoint_info_from_name(endpoint=None, one=None):
-    """
-    Extracts Globus endpoint ID and root path given a repository name that is registered in the
-    database accessed by ONE.
-
-    :param endpoint: str, repository name as registered in database
-    :param one: ONE instance, optional
-    :return: tuple of str, (endpoint_id, root_path)
-    """
-    one = one or ONE()
-    repos = one.alyx.rest('data-repository', 'list')
-    repo_names = [r['name'] for r in repos]
-    if not endpoint or endpoint not in repo_names:
-        _logger.error("Choose one of the following endpoints from the current Alyx database or "
-                      "pass an ONE instance with a different base_url:")
-        _ = [print(n) for n in repo_names]
-    else:
-        endpoint_dict = [r for r in repos if r['name'] == endpoint][0]
-        return endpoint_dict['globus_endpoint_id'], endpoint_dict['globus_path']
+            local_paths = map(Path, filter(None, path_file.read_text().strip().split(',')))
+            _logger.debug(f'Found local endpoint paths in Globus Connect settings')
+        else:
+            msg = ('Cannot find local endpoint path, check if Globus Connect is set up correctly, '
+                   '{} exists and contains a valid path.')
+            warnings.warn(msg.format(path_file))
+            local_paths = []
+        return list(local_paths)
 
 
-def get_lab_from_endpoint_id(endpoint=None, one=None):
+def get_lab_from_endpoint_id(endpoint=None, alyx=None):
     """
     Extracts lab name given an endpoint id root path given a repository name that is registered in the
     database accessed by ONE.
 
-    :param endpoint: endpoint UUID, optional if not given will get attempt to find local endpoint id
-    :param one: ONE instance, optional
-    :return: list of str, [lab name]
+    Parameters
+    ----------
+    endpoint : uuid.UUID, str
+        Endpoint UUID, optional if not given will get attempt to find local endpoint UUID.
+    alyx : one.webclient.AlyxClient
+        An instance of AlyxClient to use
+
+    Returns
+    -------
+    str
+        Lab name associated with the endpoint UUID
     """
 
-    one = one or ONE()
+    alyx = alyx or AlyxClient(silent=True)
     if not endpoint:
         endpoint = get_local_endpoint_id()
-    lab = one.alyx.rest('labs', 'list', django=f"repositories__globus_endpoint_id,{endpoint}")
+    lab = alyx.rest('labs', 'list', django=f'repositories__globus_endpoint_id,{endpoint}')
     if len(lab):
-        return [la['name'] for la in lab]
-
-
-def _remove_uuid_from_filename(file_path):
-    file_path = Path(file_path)
-    name_parts = file_path.name.split('.')
-    if len(name_parts) < 2 or not is_uuid_string(name_parts[-2]):
-        return str(file_path)
-    name_parts.pop(-2)
-    return str(file_path.parent.joinpath('.'.join(name_parts)))
+        lab_names = [la['name'] for la in lab]
+        assert len(set(lab_names)) == 1, f'Multiple labs associated with endpoint UUID {endpoint}'
+        return lab_names[0]
 
 
 def as_globus_path(path):
     """
-    Copied from ibblib.io.globus
     Convert a path into one suitable for the Globus TransferClient.  NB: If using tilda in path,
     the home folder of your Globus Connect instance must be the same as the OS home dir.
 
-    :param path: A path str or Path instance
-    :return: A formatted path string
+    Parameters
+    ----------
+    path : pathlib.Path, str
+        A path to convert to a Globus-complient path string
 
-    Examples:
+    Returns
+    -------
+    str
+        A formatted path string
+
+    Examples
+    --------
         # A Windows path
         >>> as_globus_path('E:\\FlatIron\\integration')
         >>> '/E/FlatIron/integration'
@@ -240,50 +247,58 @@ def as_globus_path(path):
     return str(path)
 
 
-class Globus:
-    """
-    Wrapper for managing files on Globus endpoints.
-    Adapted from https://github.com/int-brain-lab/ibllib/blob/ca8cd93ccac665efda6943f14047ba53503e8bb8/ibllib/io/globus.py
-    """
+class Globus(DownloadClient):
+    """Wrapper for managing files on Globus endpoints."""
     def __init__(self, client_name='default'):
         # Setting up transfer client
+        super().__init__()
         self.client = create_globus_client(client_name=client_name)
-        self.pars = iopar.read(_PAR_ID_STR + f'/{client_name}')
+        self.pars = load_client_params(f'{CLIENT_KEY}.{client_name}')
         # Try adding local endpoint
-        if hasattr(self.pars, 'local_endpoint'):
-            self.endpoints = {'local': {'id': self.pars.local_endpoint}}
-            _logger.info("Adding local endpoint.")
-            if hasattr(self.pars, 'local_path'):
-                self.endpoints['local']['root_path'] = self.pars.local_path
-        else:
-            _logger.warning("Not adding local endpoint as information is missing from parameter "
-                            "file. Add endpoints manually or run one.globus.setup() first.")
-            self.endpoints = {}
+        self.endpoints = {'local': {'id': self.pars.local_endpoint}}
+        _logger.info('Adding local endpoint.')
+        self.endpoints['local']['root_path'] = self.pars.local_path
 
-    def add_endpoint(self, endpoint, label=None, root_path=None, overwrite=False, one=None):
+    def to_address(self, data_path):
+        pass
+
+    def download_file(self, file_address):
+        pass
+
+    @staticmethod
+    def setup(*args, **kwargs):
+        pass
+
+    def add_endpoint(self, endpoint, label=None, root_path=None, overwrite=False, alyx=None):
         """
         Add an endpoint to the Globus instance to be used by other functions.
 
-        :param endpoint: UUID or str, the endpoint UUID or database repository name of the endpoint
-        :param label: str, label to access the endpoint. If endpoint is UUID this has to be set,
-        otherwise optional
-        :param root_path: str or Path, path to be accessed by globus on the endpoint.
-        :param overwrite: bool, whether existing endpoint with the same label should be replaced
-        :param one: ONE instance for extracting endpoint information from database. If endpoint is
-        UUID this has no effect.
-        :return:
+        Parameters
+        ----------
+        endpoint : uuid.UUID, str
+            The endpoint UUID or database repository name of the endpoint.
+        label : str
+            Label to access the endpoint. If endpoint is UUID this has to be set, otherwise is
+            optional.
+        root_path : str, pathlib.Path, pathlib.PurePath
+            File path to be accessed by Globus on the endpoint.
+        overwrite : bool
+            Whether existing endpoint with the same label should be replaced.
+        alyx : webclient.AlyxClient
+            An AlyxClient instance for looking up repository information.
         """
-        if is_uuid(endpoint, range(5)):
+        if is_uuid(endpoint, versions=(1,)):  # MAC address UUID
             if label is None:
-                _logger.error("If 'endpoint' is a UUID, 'label' cannot be None.")
+                raise ValueError('If "endpoint" is a UUID, "label" cannot be None.')
             endpoint_id = str(endpoint)
         else:
-            endpoint_id, globus_path = get_endpoint_info_from_name(endpoint, one=one)
-            root_path = root_path or globus_path
+            repo = self.repo_from_alyx(endpoint, alyx=alyx)
+            endpoint_id = repo['globus_endpoint_id']
+            root_path = root_path or repo['globus_path']
             label = label or endpoint
         if label in self.endpoints.keys() and overwrite is False:
-            _logger.error(f"An endpoint called '{label}' already exists. Choose a different label "
-                          f"or set overwrite=True")
+            _logger.error(f'An endpoint called "{label}" already exists. Choose a different label '
+                          'or set overwrite=True')
         else:
             self.endpoints[label] = {'id': endpoint_id}
             if root_path:
@@ -295,8 +310,8 @@ class Globus:
         if root_path and not str(path).startswith(root_path):
             path = Path(root_path).joinpath(path)
         if not path.is_absolute():
-            _logger.error("If there is no root_path for this endpoint in .endpoints, 'path' must "
-                          "be an absolute path.")
+            _logger.error('If there is no root_path for this endpoint in .endpoints, "path" must '
+                          'be an absolute path.')
         return path
 
     def _endpoint_id_root(self, endpoint):
@@ -310,10 +325,9 @@ class Globus:
             endpoint_id = endpoint
             return endpoint_id, None
         else:
-            _logger.error(f"'endpoint' must be a UUID or the label of an endpoint registered in "
-                          f"this Globus instance. You can add endpoints via the add_endpoints "
-                          f"method")
-            return
+            raise ValueError(
+                '"endpoint" must be a UUID or the label of an endpoint registered in this '
+                'Globus instance. You can add endpoints via the add_endpoints method')
 
     def ls(self, endpoint, path, remove_uuid=False, return_size=False):
         """
@@ -327,7 +341,7 @@ class Globus:
         out = []
         try:
             for entry in self.client.operation_ls(endpoint_id, path=path):
-                fn = _remove_uuid_from_filename(entry['name']) if remove_uuid else entry['name']
+                fn = remove_uuid_file(entry['name'], dry=True) if remove_uuid else entry['name']
                 if return_size:
                     size = entry['size'] if entry['type'] == 'file' else None
                     out.append((fn, size))
@@ -367,7 +381,7 @@ class Globus:
         """
         try:
             task_id = globus_func()
-            print(f"Waiting for Globus task {task_id} to complete")
+            print(f'Waiting for Globus task {task_id} to complete')
             # While the task with task is active, print a dot every second. Timeout after timeout
             i = 0
             while not self.client.task_wait(task_id, timeout=5, polling_interval=1):
@@ -375,8 +389,8 @@ class Globus:
                 i += 1
                 if time_out and i >= time_out:
                     task = self.client.get_task(task_id)
-                    raise IOError(f"Globus task {task_id} timed out after {time_out} seconds, "
-                                  f"with task status {task['status']}")
+                    raise IOError(f'Globus task {task_id} timed out after {time_out} seconds, '
+                                  f'with task status {task["status"]}')
             task = self.client.get_task(task_id)
             if task['status'] == 'SUCCEEDED':
                 # Sometime Globus sets the status to SUCCEEDED but doesn't truly finish.
@@ -384,24 +398,24 @@ class Globus:
                 try:
                     successful = self.client.task_successful_transfers(task_id)
                     skipped = self.client.task_skipped_errors(task_id)
-                    print(f"\nGlobus task {task_id} completed."
-                          f"\nSkipped transfers: {len(list(skipped))}"
-                          f"\nSuccessful transfers: {len(list(successful))}")
+                    print(f'\nGlobus task {task_id} completed.'
+                          f'\nSkipped transfers: {len(list(skipped))}'
+                          f'\nSuccessful transfers: {len(list(successful))}')
                     for info in successful:
-                        _logger.debug(f"{info['source_path']} -> {info['destination_path']}")
+                        _logger.debug(f'{info["source_path"]} -> {info["destination_path"]}')
                 except TransferAPIError:
-                    _logger.warning(f"\nGlobus task {task_id} SUCCEEDED but querying transfers was"
-                                    f"unsuccessful")
+                    _logger.warning(f'\nGlobus task {task_id} SUCCEEDED but querying transfers was'
+                                    f'unsuccessful')
             else:
-                raise IOError(f"Globus task finished unsucessfully with status {task['status']}")
+                raise IOError(f'Globus task finished unsucessfully with status {task["status"]}')
             return task_id
         except (GlobusAPIError, NetworkError, GlobusTimeoutError, GlobusConnectionError,
                 GlobusConnectionTimeoutError) as e:
             if retries < 1:
-                _logger.error(f"\nRetried too many times.")
+                _logger.error(f'\nRetried too many times.')
                 raise e
             else:
-                _logger.debug("\nGlobus experienced a network error", exc_info=True)
+                _logger.debug('\nGlobus experienced a network error', exc_info=True)
                 # if we reach this point without returning or erroring, retry
-                _logger.warning("\nGlobus experienced a network error, retrying.")
+                _logger.warning('\nGlobus experienced a network error, retrying.')
                 self.run_task(globus_func, retries=(retries - 1), time_out=time_out)