+++ CHANGELOG.md
@@ -1,11 +1,16 @@
 # Changelog
+## [Latest](https://github.com/int-brain-lab/ONE/commits/main) [1.6.2]
+
+### Modified
+
+- more readable error message; raw JSON printed to debug logger
+
 ## [Latest](https://github.com/int-brain-lab/ONE/commits/main) [1.6.1]
 
 ### Modified
 
 - rest command loging includes the whole json field on error
 - added silent option to instantiate One on local files
-
 
 ## [1.6.0]